% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juggle.R
\name{juggle}
\alias{juggle}
\title{Juggle JAGS Model Code}
\usage{
juggle(model_code)
}
\arguments{
\item{model_code}{A string of the JAGS model to manipulate.}
}
\value{
A jg_model object.
}
\description{
Juggle JAGS Model Code
}

